Release: OpenWrap_Q1_28-Mar-23

+++ endpoints/openrtb2/ctv/constant/constant.go
@@ -27,7 +27,7 @@ var (
 )
 
 // BidStatus contains bids filtering reason
-type BidStatus = int
+type BidStatus = int64
 
 const (
 	//StatusOK ...
